+++ test/cookbooks/docker_test/files/default/tests/minitest/binary_test.rb
@@ -2,5 +2,4 @@ require File.expand_path('../support/helpers', __FILE__)
 
 describe_recipe "docker_test::binary" do
   include Helpers::DockerTest
-  
 end