+++ test/cookbooks/docker_test/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file is used to list changes made in each version of docker_test.
 
+## 0.4.0:
+
+* Bugfix: Remove deprecated public_port in container_lwrp
+* Bugfix: Add `init_type false` for busybox test containers
+* Enhancement: Add bflad/testcontainerd image, container, and tests
+
 ## 0.3.0:
 
 * Enhancement: Change Dockerfile FROM to already downloaded busybox image instead of ubuntu